ai-serving/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
tensorflow